Refactor component into single directory
