+++ switch.py
@@ -4,10 +4,10 @@ https://home-assistant.io/components/binary_sensor.kuna/
 """
 import logging
 
-from homeassistant.components.binary_sensor import BinarySensorDevice
+from homeassistant.components.switch import SwitchDevice
+from . import DOMAIN
 
 DEPENDENCIES = ['kuna']
-DOMAIN = 'kuna'
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -19,52 +19,46 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
     devices = []
 
     for camera in kuna.account.cameras:
-        device = KunaBinarySensor(kuna, camera)
+        device = KunaSwitch(kuna, camera)
         devices.append(device)
 
     add_entities(devices)
 
 
-class KunaBinarySensor(BinarySensorDevice):
+class KunaSwitch(SwitchDevice):
 
     def __init__(self, kuna, camera):
         self._account = kuna
         self._camera = camera
         self._original_id = self._camera.serial_number
-        self._name = '{} Motion'.format(self._camera.name)
-        self._unique_id = '{}-Motion'.format(self._camera.serial_number)
-        self._device_class = 'motion'
-
-    @property
-    def should_poll(self):
-        return False
+        self._name = '{} Switch'.format(self._camera.name)
+        self._unique_id = '{}-Switch'.format(self._camera.serial_number)
 
     @property
     def available(self):
         return self._camera.status
 
     @property
+    def should_poll(self):
+        return False
+
+    @property
     def unique_id(self):
-        """Return the unique ID of the binary sensor."""
+        """Return the unique ID of the switch."""
         return self._unique_id
 
     @property
     def name(self):
-        """Return the name of the binary sensor."""
+        """Return the name of the switch."""
         return self._name
 
     @property
-    def device_class(self):
-        """Return the class of the binary sensor."""
-        return self._device_class
-
-    @property
     def is_on(self):
-        """Return the state of the binary sensor."""
-        return self._camera.recording_active
+        """Return the state of the switch."""
+        return self._camera.bulb_on
 
     def update(self):
-        """Read new state data from the library."""
+        """Read new state data from the updated account camera list."""
         kuna = self._account
         for camera in kuna.account.cameras:
             if camera.serial_number == self._original_id:
@@ -77,3 +71,13 @@ class KunaBinarySensor(BinarySensorDevice):
     async def async_added_to_hass(self):
         """Add callback after being added to hass."""
         self._account.add_update_listener(self.update_callback)
+
+    def turn_on(self, **kwargs):
+        """Turn the switch on."""
+        self._camera.light_on()
+        self._account.update()
+
+    def turn_off(self, **kwargs):
+        """Turn the device off."""
+        self._camera.light_off()
+        self._account.update()
